Change file names & make lists of test data
