+++ .github/workflows/firebase-ci.yaml
@@ -14,7 +14,7 @@ jobs:
         run: npm install -g firebase-tools
       - name: Install Jest
         run: npm install --save-dev jest
-      - name: Install playwright
+      - name: Install Playwright
         run: npx playwright install
       - name: Install dependencies
         run: npm install
